renamed template controller to page controller, moved facebook apis to facebookapi helper
