+++ app/views/page/gallery.html.erb
