+++ app/views/page/home.html.erb
