+++ app/views/page/news.html.erb
