+++ app/views/page/photos.html.erb
